google cloud properly fixed
