+++ back-end/src/analysis/GoogleTextToSpeechComponent.ts
@@ -1,0 +1,36 @@
+
+import { AnalysisComponent } from "./AnalysisComponent"
+
+const speech : any = require('@google-cloud/speech');
+
+class GoogleTextToSpeechComponent extends AnalysisComponent<string,string>{
+  speechClient = new speech.SpeechClient();
+  inputTopic = "audioBytes";
+  outputTopic = "transcript";
+
+  analyze(audioBytes : string) : Promise<string>{
+    return new Promise((resolve,reject) => {
+      const audio = {
+        content: audioBytes,
+      };
+      const config = {
+        languageCode: 'en-US',
+        audio_channel_count: 1,
+        enableAutomaticPunctuation: true,
+      };
+      const request = {
+        audio: audio,
+        config: config,
+      };
+      this.speechClient.recognize(request).then(response => {
+        var [transcriptParts] = response
+        const transcription = transcriptParts.results
+          .map(result => result.alternatives[0].transcript)
+          .join('\n');
+        resolve(transcription)
+      })
+    });
+  }
+}
+
+export default GoogleTextToSpeechComponent;
